Sumble files
Sumble is a numeric game. Players have six attempts to guess a 5-digit numeric code, where the sum of the digits must equal a target number. Feedback is given for each guess in the form of coloured tiles indicating when the digits match or occupy the correct position. It is built with MATLAB® and inspired by the web-based word game, Wordle, published by The New York Times Company™.